move Signer into primitives
