+++ mod.ts
@@ -4,7 +4,6 @@ export * as Z from "./deps/zones.ts"
 export * from "./effects/mod.ts"
 export * as fluent from "./fluent/mod.ts"
 export * as frame from "./frame_metadata/mod.ts"
-export { type Signer } from "./frame_metadata/mod.ts"
 export * as ink from "./ink_metadata/mod.ts"
 export * from "./primitives/mod.ts"
 export * as rpc from "./rpc/mod.ts"
