+++ primitives/mod.ts
@@ -1,2 +1,3 @@
 export * from "./MultiAddress.ts"
 export * from "./MultiSignature.ts"
+export * from "./Signer.ts"
